Refactored the use of Entities
