+++ src/database/index.ts
@@ -1,0 +1,6 @@
+import { Discussions } from './entity/channelDiscussion';
+import { Channels } from './entity/channels';
+import { Hashtags } from './entity/hashtag';
+import { Users } from './entity/user';
+
+export { Discussions, Channels, Hashtags, Users };
