+++ src/modules/User/user.repository.ts
@@ -1,6 +1,12 @@
+/**
+ * @description Refactored the import of Users
+ *
+ * @author Shreyash <pandeyshreyash2201@gmail.com>
+ */
+
 import { Injectable } from '@nestjs/common';
 import { EntityRepository, Repository } from 'typeorm';
-import { Users } from '../../database/entity/user';
+import { Users } from '../../database';
 
 @EntityRepository(Users)
 export class UserRepository extends Repository<Users> {}
