+++ src/modules/channelDiscussions/discussion.service.ts
@@ -1,18 +1,17 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 /**
- * Created Discussion Service
+ * @description Refactored the import of Discussions
  *
- * @author Devesh 
+ * @author Shreyash <pandeyshreyash2201@gmail.com>
  */
 
 import { getCustomRepository } from 'typeorm';
-import { Discussions } from '../../database/entity/channelDiscussion';
+import { Discussions } from '../../database';
 import { Injectable, NotFoundException, InternalServerErrorException } from '@nestjs/common';
 import { Transactional } from 'typeorm-transactional-cls-hooked';
 import { DiscussionRepository } from './discussion.repository';
 import { RedisService } from '../../shared/services/redis.service';
 import { UpdateDiscussionDto } from './dto';
-
 
 @Injectable()
 export class DiscussionService {
@@ -34,7 +33,6 @@ export class DiscussionService {
 		});
 	}
 
-
 	@Transactional()
 	async findById(DiscussionId: number): Promise<Discussions> {
 		const Discussion = await this.DiscussionRepo.findOne({
@@ -48,37 +46,31 @@ export class DiscussionService {
 	}
 
 	@Transactional()
-	async updateDiscussion(
-		DiscussionId: number,
-		Discussion: UpdateDiscussionDto
-	): Promise<Discussions | undefined> {
-		try{
+	async updateDiscussion(DiscussionId: number, Discussion: UpdateDiscussionDto): Promise<Discussions | undefined> {
+		try {
 			const postRes = await this.DiscussionRepo.update(DiscussionId, Discussion);
-			if (postRes.affected===0) {
+			if (postRes.affected === 0) {
 				throw new NotFoundException();
 			}
 			return await this.findById(DiscussionId);
+		} catch (e) {
+			throw e;
 		}
-		catch(e){
-			throw e;;
-		}	
 	}
 
 	@Transactional()
 	async deleteDiscussion(DiscussionId: number): Promise<Object | undefined> {
-		try{
+		try {
 			const delRes = await this.DiscussionRepo.delete({ id: DiscussionId });
-			if (delRes.affected===0) {
+			if (delRes.affected === 0) {
 				throw new NotFoundException();
+			} else {
+				return {
+					message: `Discussion with id ${DiscussionId} has been deleted`
+				};
 			}
-			else{
-				return{
-					message:`Discussion with id ${DiscussionId} has been deleted`
-				}							
-			}
+		} catch (e) {
+			throw e;
 		}
-		catch(e){
-			throw e;;
-		}		
 	}
 }
