+++ src/modules/channelDiscussions/dto/update-discussion.dto.ts
@@ -2,7 +2,7 @@
 /**
  * Created DTO
  *
- * @author Devesh 
+ * @author Devesh
  */
 
 export class UpdateDiscussionDto {
